requirements.txt
attrs==17.3.0
bleach==1.5.0
certifi==2017.11.5
chardet==3.0.4
cycler==0.10.0
enum34==1.1.6
html5lib==0.9999999
idna==2.6
Keras==2.1.2
Markdown==2.6.10
matplotlib==2.1.1
numpy==1.14.0rc1
pandas==0.21.1
pandas-datareader==0.5.0
protobuf==3.5.0.post1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
requests-file==1.4.2
requests-ftp==0.3.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorflow==1.4.1
tensorflow-tensorboard==0.4.0rc3
urllib3==1.22
Werkzeug==0.13
